% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbov.R
\name{look_up}
\alias{look_up}
\title{Looks up the value stored in the given gbov by the specified hash
and returns it if it was found, otherwise NULL is returned
#' @export}
\usage{
look_up(gbov, hash)
}
\description{
Looks up the value stored in the given gbov by the specified hash
and returns it if it was found, otherwise NULL is returned
#' @export
}
