% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbov.R
\name{save.gbov}
\alias{save.gbov}
\title{By default takes a gbov object and saves it in the given path with
the specified name and returns the path to the saved file
if new is FALSE, it overwrites gbov.RDS in the given directory(path)
#' @export}
\usage{
save.gbov(gbov, path = ".", name = "new-gbov", new = TRUE)
}
\description{
By default takes a gbov object and saves it in the given path with
the specified name and returns the path to the saved file
if new is FALSE, it overwrites gbov.RDS in the given directory(path)
#' @export
}
